Split file from nat.py for extending another nat features
