+++ acos_client/v30/nat/__init__.py
@@ -1,0 +1,25 @@
+# Copyright 2015, A10 Networks
+#
+#    Licensed under the Apache License, Version 2.0 (the "License"); you may
+#    not use this file except in compliance with the License. You may obtain
+#    a copy of the License at
+#
+#         http://www.apache.org/licenses/LICENSE-2.0
+#
+#    Unless required by applicable law or agreed to in writing, software
+#    distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+#    WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+#    License for the specific language governing permissions and limitations
+#    under the License.
+from __future__ import absolute_import
+from __future__ import unicode_literals
+
+from acos_client import errors as acos_errors
+from acos_client.v30 import base
+from acos_client.v30.nat.pool import Pool
+
+
+class Nat(base.BaseV30):
+    @property
+    def pool(self):
+        return Pool(self.client)
